i64

+++ include/tts/tspans/constants.hpp
@@ -6,32 +6,32 @@
 
 namespace tts::timespan {
     // constants!
-    constexpr i32 SECONDS{1};
+    constexpr i64 SECONDS{1};
 
         /**
          * Number of seconds in a second. Base (reference) unit 
          * for timespans.
         */
 
-    constexpr i32 MINUTES{60};
+    constexpr i64 MINUTES{60};
         
         /**
          * Number of seconds in a minute.
         */   
  
-    constexpr i32 HOURS{60 * 60};
+    constexpr i64 HOURS{60 * 60};
         
         /**
          * Number of seconds in an hour.
         */   
  
-    constexpr i32 DAYS{24 * 60 * 60};
+    constexpr i64 DAYS{24 * 60 * 60};
         
         /**
          * Number of seconds in a day.
         */   
  
-    constexpr i32 WEEKS{7 * 24 * 60 * 60};
+    constexpr i64 WEEKS{7 * 24 * 60 * 60};
         
         /**
          * Number of seconds in a week.
